+++ csharp/ToolGood.FlowVision/ToolGood.Algorithm2/MyDate.cs
@@ -93,7 +93,7 @@ namespace ToolGood.Algorithm2
 		/// 构造函数
 		/// </summary>
 		/// <param name="num"></param>
-		public MyDate(double num)
+		public MyDate(decimal num)
 		{
 			int days = (int)num;
 			if (days > 365) {
@@ -106,10 +106,10 @@ namespace ToolGood.Algorithm2
 				Month = null;
 				Day = days;
 			}
-			double d = num - days;
+            decimal d = num - days;
 			Hour = (int)(d * 24);
-			Minute = (int)((d * 24 - Hour) * 60.0);
-			Second = (int)(((d * 24 - Hour) * 60.0 - Minute) * 60.0);
+			Minute = (int)((d * 24 - Hour) * 60.0m);
+			Second = (int)(((d * 24 - Hour) * 60.0m - Minute) * 60.0m);
 		}
 
 		/// <summary>
@@ -344,28 +344,28 @@ namespace ToolGood.Algorithm2
 		}
 
 		/// <summary>
-		/// double=>MyDate
+		/// decimal=>MyDate
 		/// </summary>
 		/// <param name="days"></param>
-		public static implicit operator MyDate(double days)
+		public static implicit operator MyDate(decimal days)
 		{
 			return new MyDate(days);
 		}
 
 		/// <summary>
-		/// MyDate=>double
+		/// MyDate=>decimal
 		/// </summary>
 		/// <param name="MyDate"></param>
-		public static implicit operator double(MyDate MyDate)
+		public static implicit operator decimal(MyDate MyDate)
 		{
 			if (MyDate.Year != null && MyDate.Year > 1900) {
 				var dt = new DateTime((MyDate.Year ?? 0), (MyDate.Month ?? 0), (MyDate.Day ?? 0), MyDate.Hour, MyDate.Minute, MyDate.Second);
-				//double days = (double)(dt -  DateTime.MinValue).TotalDays;
-				double days = (double)(dt - new DateTime(1900, 1, 1)).TotalDays + 2;
-				days += (MyDate.Hour + (MyDate.Minute + MyDate.Second / 60.0) / 60) / 24;
+				//decimal days = (decimal)(dt -  DateTime.MinValue).TotalDays;
+				decimal days = (decimal)(dt - new DateTime(1900, 1, 1)).TotalDays + 2;
+				days += (MyDate.Hour + (MyDate.Minute + MyDate.Second / 60.0m) / 60) / 24;
 				return days;
 			}
-			return (MyDate.Day ?? 0) + (MyDate.Hour + (MyDate.Minute + MyDate.Second / 60.0) / 60) / 24;
+			return (MyDate.Day ?? 0) + (MyDate.Hour + (MyDate.Minute + MyDate.Second / 60.0m) / 60) / 24;
 		}
 
 		/// <summary>
@@ -381,7 +381,7 @@ namespace ToolGood.Algorithm2
 			} else if (myDate.Year != null && num.Year == null) {
 				return new MyDate(num.ToDateTime().Add(myDate.ToTimeSpan()));
 			}
-			return (MyDate)((double)myDate + (double)num);
+			return (MyDate)((decimal)myDate + (decimal)num);
 		}
 
 		/// <summary>
@@ -397,7 +397,7 @@ namespace ToolGood.Algorithm2
 			} else if (myDate.Year != null && num.Year == null) {
 				return new MyDate(num.ToDateTime().Add(myDate.ToTimeSpan().Negate()));
 			}
-			return (MyDate)((double)myDate - (double)num);
+			return (MyDate)((decimal)myDate - (decimal)num);
 		}
 
 		/// <summary>加
@@ -406,9 +406,9 @@ namespace ToolGood.Algorithm2
 		/// <param name="myDate"></param>
 		/// <param name="num"></param>
 		/// <returns></returns>
-		public static MyDate operator +(MyDate myDate, double num)
+		public static MyDate operator +(MyDate myDate, decimal num)
 		{
-			return (MyDate)((double)myDate + (double)num);
+			return (MyDate)((decimal)myDate + (decimal)num);
 		}
 
 		/// <summary>
@@ -417,9 +417,9 @@ namespace ToolGood.Algorithm2
 		/// <param name="myDate"></param>
 		/// <param name="num"></param>
 		/// <returns></returns>
-		public static MyDate operator -(MyDate myDate, double num)
+		public static MyDate operator -(MyDate myDate, decimal num)
 		{
-			return (MyDate)((double)myDate - (double)num);
+			return (MyDate)((decimal)myDate - (decimal)num);
 		}
 
 		/// <summary>
@@ -428,14 +428,14 @@ namespace ToolGood.Algorithm2
 		/// <param name="myDate"></param>
 		/// <param name="num"></param>
 		/// <returns></returns>
-		public static MyDate operator *(MyDate myDate, double num)
+		public static MyDate operator *(MyDate myDate, decimal num)
 		{
 #if NETSTANDARD2_1
-			if (myDate.Year != null) {
-				return new MyDate(myDate.ToTimeSpan().Multiply(num));
-			}
+            if (myDate.Year != null) {
+                return new MyDate(myDate.ToTimeSpan().Multiply((double)num));
+            }
 #endif
-			return (MyDate)((double)myDate * (double)num);
+			return (MyDate)((decimal)myDate * (decimal)num);
 		}
 
 		/// <summary>
@@ -444,14 +444,14 @@ namespace ToolGood.Algorithm2
 		/// <param name="myDate"></param>
 		/// <param name="num"></param>
 		/// <returns></returns>
-		public static MyDate operator /(MyDate myDate, double num)
+		public static MyDate operator /(MyDate myDate, decimal num)
 		{
 #if NETSTANDARD2_1
-			if (myDate.Year != null) {
-				return new MyDate(myDate.ToTimeSpan().Divide(num));
-			}
+            if (myDate.Year != null) {
+                return new MyDate(myDate.ToTimeSpan().Divide((double)num));
+            }
 #endif
-			return (MyDate)((double)myDate / (double)num);
+            return (MyDate)((decimal)myDate / (decimal)num);
 		}
 
 		#endregion operator
